changed example project name
