+++ Akkatecture.sln
@@ -31,10 +31,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Akkatecture.Examples.Seed",
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Akkatecture.Examples.Worker", "examples\cluster\Akkatecture.Examples.Worker\Akkatecture.Examples.Worker.csproj", "{DD833121-B2AD-4F2E-A12E-B1B28714817A}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Akkatecture.Examples.UserAccount", "examples\simple\Akkatecture.Examples.UserAccount\Akkatecture.Examples.UserAccount.csproj", "{8D9E347F-26AA-40E6-931B-763DE99F2B01}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Akkatecture.Examples.UserAccount.Application", "examples\simple\Akkatecture.Examples.UserAccount.Application\Akkatecture.Examples.UserAccount.Application.csproj", "{2D58822C-DE63-43C6-A50A-384128D22548}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Akkatecture.Examples.ClusterClient", "examples\cluster\Akkatecture.Examples.ClusterClient\Akkatecture.Examples.ClusterClient.csproj", "{5513AA4C-5E65-41CC-84AF-6C6003B54F60}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "walkthrough", "walkthrough", "{114EF159-E6C3-43AC-A6C4-7248AC39B6AE}"
@@ -50,6 +46,10 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "build", "build", "{8C65E1C0
 ProjectSection(SolutionItems) = preProject
 	appveyor.yml = appveyor.yml
 EndProjectSection
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Akkatecture.Examples.Application", "examples\simple\Akkatecture.Examples.Application\Akkatecture.Examples.Application.csproj", "{F4512D84-8576-44DB-9F15-E299F62EB9D3}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Akkatecture.Examples.Domain", "examples\simple\Akkatecture.Examples.Domain\Akkatecture.Examples.Domain.csproj", "{C104C065-D0CE-42C0-9C43-D8F17ECEA0A0}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -81,14 +81,6 @@ Global
 		{DD833121-B2AD-4F2E-A12E-B1B28714817A}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{DD833121-B2AD-4F2E-A12E-B1B28714817A}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{DD833121-B2AD-4F2E-A12E-B1B28714817A}.Release|Any CPU.Build.0 = Release|Any CPU
-		{8D9E347F-26AA-40E6-931B-763DE99F2B01}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{8D9E347F-26AA-40E6-931B-763DE99F2B01}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{8D9E347F-26AA-40E6-931B-763DE99F2B01}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{8D9E347F-26AA-40E6-931B-763DE99F2B01}.Release|Any CPU.Build.0 = Release|Any CPU
-		{2D58822C-DE63-43C6-A50A-384128D22548}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{2D58822C-DE63-43C6-A50A-384128D22548}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{2D58822C-DE63-43C6-A50A-384128D22548}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{2D58822C-DE63-43C6-A50A-384128D22548}.Release|Any CPU.Build.0 = Release|Any CPU
 		{5513AA4C-5E65-41CC-84AF-6C6003B54F60}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{5513AA4C-5E65-41CC-84AF-6C6003B54F60}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{5513AA4C-5E65-41CC-84AF-6C6003B54F60}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -101,6 +93,14 @@ Global
 		{1349C01A-9C7F-48FA-A651-2C24F5E8A130}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{1349C01A-9C7F-48FA-A651-2C24F5E8A130}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{1349C01A-9C7F-48FA-A651-2C24F5E8A130}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F4512D84-8576-44DB-9F15-E299F62EB9D3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F4512D84-8576-44DB-9F15-E299F62EB9D3}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F4512D84-8576-44DB-9F15-E299F62EB9D3}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F4512D84-8576-44DB-9F15-E299F62EB9D3}.Release|Any CPU.Build.0 = Release|Any CPU
+		{C104C065-D0CE-42C0-9C43-D8F17ECEA0A0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C104C065-D0CE-42C0-9C43-D8F17ECEA0A0}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C104C065-D0CE-42C0-9C43-D8F17ECEA0A0}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C104C065-D0CE-42C0-9C43-D8F17ECEA0A0}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -114,12 +114,12 @@ Global
 		{F39AA664-CAE8-4592-9173-8D03CABA1C07} = {7522402D-176E-4305-A68B-F3FFBFD38158}
 		{BEAC89B6-9B71-4BF6-9B3A-4193DE028F7A} = {F39AA664-CAE8-4592-9173-8D03CABA1C07}
 		{DD833121-B2AD-4F2E-A12E-B1B28714817A} = {F39AA664-CAE8-4592-9173-8D03CABA1C07}
-		{8D9E347F-26AA-40E6-931B-763DE99F2B01} = {090ECC2F-5647-4B15-ABDF-2E1B44439626}
-		{2D58822C-DE63-43C6-A50A-384128D22548} = {090ECC2F-5647-4B15-ABDF-2E1B44439626}
 		{5513AA4C-5E65-41CC-84AF-6C6003B54F60} = {F39AA664-CAE8-4592-9173-8D03CABA1C07}
 		{114EF159-E6C3-43AC-A6C4-7248AC39B6AE} = {7522402D-176E-4305-A68B-F3FFBFD38158}
 		{B9F45061-8C53-4F90-ACC3-6A03728F5D1E} = {114EF159-E6C3-43AC-A6C4-7248AC39B6AE}
 		{1349C01A-9C7F-48FA-A651-2C24F5E8A130} = {114EF159-E6C3-43AC-A6C4-7248AC39B6AE}
+		{F4512D84-8576-44DB-9F15-E299F62EB9D3} = {090ECC2F-5647-4B15-ABDF-2E1B44439626}
+		{C104C065-D0CE-42C0-9C43-D8F17ECEA0A0} = {090ECC2F-5647-4B15-ABDF-2E1B44439626}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {F5A40806-ED71-4286-A37C-BE0364FA29A6}
