+++ examples/simple/Akkatecture.Examples.Application/Program.cs
@@ -21,19 +21,37 @@
 // IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
 // CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
-using Akkatecture.Core;
-using Akkatecture.ValueObjects;
-using Newtonsoft.Json;
+using System;
+using Akka.Actor;
+using Akkatecture.Examples.Domain.Model.UserAccount;
+using Akkatecture.Examples.Domain.Model.UserAccount.Commands;
 
-namespace Akkatecture.Examples.UserAccount.Domain.UserAccountModel
+namespace Akkatecture.Examples.Application
 {
-    [JsonConverter(typeof(SingleValueObjectConverter))]
-    public class UserAccountId : Identity<UserAccountId>
+    public class Program
     {
-        public UserAccountId(string value)
-            : base(value)
+        public static void Main(string[] args)
         {
+            //Create actor system
+            var system = ActorSystem.Create("useraccount-example");
+
+            //Create supervising aggregate manager for UserAccount aggregate root actors
+            var aggregateManager = system.ActorOf(Props.Create(() => new UserAccountAggregateManager()));
+
+            //Build create user account aggregate command
+            var aggregateId = UserAccountId.New;
+            var createUserAccountCommand = new CreateUserAccountCommand(aggregateId, "foo bar");
             
+            //Send command, this is equivalent to command.publish() in other cqrs frameworks
+            aggregateManager.Tell(createUserAccountCommand);
+            
+            var changeNameCommand = new UserAccountChangeNameCommand(aggregateId, "foo bar baz");
+            aggregateManager.Tell(changeNameCommand);
+                        
+            //block end of program
+            Console.ReadLine();
         }
+
+       
     }
 }
