+++ examples/simple/Akkatecture.Examples.Domain/Akkatecture.Examples.Domain.csproj
@@ -1,29 +1,28 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <OutputType>Exe</OutputType>
     <TargetFramework>netcoreapp2.0</TargetFramework>
     <TreatWarningsAsErrors>True</TreatWarningsAsErrors>
     <GenerateAssemblyInfo>False</GenerateAssemblyInfo>
     <GeneratePackageOnBuild>False</GeneratePackageOnBuild>
     <PackageRequireLicenseAcceptance>True</PackageRequireLicenseAcceptance>
-    <Title>Akkatecture.Examples.UserAccount.Application</Title>
+    <Title>Akkatecture.Examples.Domain</Title>
     <Authors>Lutando Ngqakaza</Authors>
     <Company>Lutando Ngqakaza</Company>
     <Copyright>Copyright (c) Lutando Ngqakaza 2018 - 2018</Copyright>
     <Description>
-      A simple "Hello World" example of how to use Akkatecture.
+      The domain project for the simple Akkatecture "Hello world" example
     </Description>
     <PackageTags>akka cqrs es eventsourcing</PackageTags>
     <RepositoryType>git</RepositoryType>
     <RepositoryUrl>https://github.com/Lutando/Akkatecture</RepositoryUrl>
-    <PackageLicenseUrl>https://raw.githubusercontent.com/Lutando/Akkatecture/master/LICENSE</PackageLicenseUrl>
+      <PackageLicenseUrl>https://raw.githubusercontent.com/Lutando/Akkatecture/master/LICENSE</PackageLicenseUrl>
     <NeutralLanguage>en-GB</NeutralLanguage>
     <PackageReleaseNotes>Externally Updated</PackageReleaseNotes>
   </PropertyGroup>
-  
+
   <ItemGroup>
-    <ProjectReference Include="..\Akkatecture.Examples.UserAccount\Akkatecture.Examples.UserAccount.csproj" />
+    <ProjectReference Include="..\..\..\src\Akkatecture\Akkatecture.csproj" />
   </ItemGroup>
 
 </Project>
