+++ examples/simple/README.md
@@ -2,7 +2,7 @@
 
 The akkatecture simple sample is the most "hello world" sample of Akkatecture. what will follow is an explanation of the two projects that relate to this example.
 
-# Akkatecture.Examples.UserAccount
+# Akkatecture.Examples.Domain
 
 This is the domain of the sample which is a basic model of user accounts, it supports two intents/commands, the creation of a user account and you can also change the user account's name. These commands if successful result in domain events being emitted that then are applied to the aggregates state, persisted to the event store, and are then published  through the actor system as domain events.
 
@@ -23,4 +23,4 @@ In this sample, we instantiate the actor system and the various domain entities 
 initialize this, is the `UserAccountAggregateManager`. We then interface with the domain by telling the manager to create user accounts 
 by instantiating and telling it a `CreateUserAccountCommand`.
 
-> to run the application in jetbrains rider or visual studio code, run the `Akkatecture.Examples.UserAccount.Application` configuration in the IDE.
+> to run the application in jetbrains rider or visual studio code, run the `Akkatecture.Examples.Application` configuration in the IDE.
