+++ examples/walkthrough/Akkatecture.Walkthrough.Application/Program.cs
@@ -22,7 +22,6 @@
 // CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
 using System;
-using System.Security.Cryptography.X509Certificates;
 using System.Threading.Tasks;
 using Akka.Actor;
 using Akkatecture.Walkthrough.Domain.Model.Account;
