+++ examples/walkthrough/Akkatecture.Walkthrough.Domain/Sagas/MoneyTransfer/MoneyTransferSagaManager.cs
@@ -23,7 +23,6 @@
 
 using System;
 using System.Linq.Expressions;
-using Akkatecture.Aggregates;
 using Akkatecture.Sagas.AggregateSaga;
 
 namespace Akkatecture.Walkthrough.Domain.Sagas.MoneyTransfer
