+++ src/Akkatecture/Aggregates/AggregateManager.cs
@@ -23,7 +23,6 @@
 
 using System;
 using Akka.Actor;
-using Akka.Configuration;
 using Akka.Event;
 using Akka.Persistence;
 using Akkatecture.Commands;
