+++ test/Akkatecture.Tests/UnitTests/Subscribers/SubscriberTests.cs
@@ -27,7 +27,6 @@ using Akka.TestKit.Xunit2;
 using Akkatecture.TestHelpers.Aggregates;
 using Akkatecture.TestHelpers.Aggregates.Commands;
 using Akkatecture.TestHelpers.Aggregates.Events;
-using Akkatecture.TestHelpers.Akka;
 using Akkatecture.TestHelpers.Subscribers;
 using Xunit;
